Refactor: #21 auth와 user 분리
