+++ src/controllers/userController.ts
@@ -49,69 +49,6 @@ const userController = {
     res.status(200).end();
   },
 
-  // 이메일 인증번호 보내기
-  async getAuthCode(req: Request, res: Response) {
-    try {
-      await userService.SendEmail(req.body.email);
-      res.status(200).end();
-    } catch (err: unknown) {
-      if (err instanceof Error) {
-        res.status(400).json(err.message);
-      }
-    }
-  },
-
-  // 인증번호 체크하기
-  async checkAuthCode(req: Request, res: Response) {
-    const { email, authCode } = req.body.data;
-    try {
-      await userService.CheckAuthMail({ email, authCode });
-      res.status(200).end();
-    } catch (err: unknown) {
-      if (err instanceof Error) {
-        res.status(400).json(err.message);
-      }
-    }
-  },
-
-  //회원가입하기
-  async registerUser(req: Request, res: Response) {
-    try {
-      const newUser = await userService.SignUp(req.body);
-      res.status(200).json(newUser);
-    } catch (err: unknown) {
-      if (err instanceof Error) {
-        res.status(400).json(err.message);
-      }
-    }
-  },
-
-  //로그인
-  async loginUser(req: Request, res: Response) {
-    try {
-      const { user, token } = await userService.SignIn(req.body.data);
-
-      if (!user) throw new Error("undefined user");
-
-      res.cookie("accessToken", token.accessToken, {
-        httpOnly: true,
-        sameSite: "none",
-        secure: true,
-      });
-      res.cookie("refreshToken", token.refreshToken, {
-        httpOnly: true,
-        sameSite: "none",
-        secure: true,
-      });
-
-      res.status(200).json(user);
-    } catch (err: unknown) {
-      if (err instanceof Error) {
-        res.status(400).json(err!.message);
-      }
-    }
-  },
-
   async registerWebPushToken(
     req: RequestHasBody<{ token: string }>,
     res: Response,
